[IMA-7645] Change JSON parsing to Swift Codable (#1)
Change Description: These changes remove the original Mantle & Cocoapods dependency of the project & change towards parsing using Swift Codable. Not only will this make all the Swift developers happy, but also we can't put a dynamic framework into a macOS command line tool.

The class declarations are based off of the output of "xcrun xcresulttool formatDescription". It does appear from initial testing that when Apple declares a property as non-optional, they don't mean that it'll be declared in the JSON for sure as many non-optional properties we found were not declared in XCResult jsons we extracted & so would fail Swift Codable parsing unless we were more relaxed in parsing & if those values weren't present, set the type to the default value (booleans to false, strings to empty string, arrays to empty arrays, etc.)

No effort has been done to ensure the parsed data structures can be coded back into a JSON object identical to what we received (or even that you can code them back). Left as exercise for the (code) reader for now.

We should also likely go back and clean up all the throwing we're doing.