Fixes #10 xcparse doesn't support space in file names & paths (#11)
Change Description: These changes make it so that our calls to xcresulttool & other command line tools will properly escape the file path in order to support paths with spaces.

Test Plan/Testing Performed: Created a folder in /tmp with xcresults named with spaces as well as destination paths that included spaces in the path & ran xcparse. Additionally tried the interactive paths. Need to add a better testing framework with some sample xcresults so this can be automated.