Fixes #16 subtests keyNotFound in ActionTestSummaryGroup (#17)
Change Description: These changes address the strict parsing issue with non-optional arrays #16 . If the JSON from xcresulttool did not return a key/value for the non-optional arrays, we would end up crashing with Illegal Instruction 4 keyNotFound as we were requiring to find that key. These changes make it so that for these non-optional arrays, we see if the key is present in the JSON & use it if so. If it isn't, we fill in with an empty array.

Test Plan/Testing Performed: Ran against SwiftiumTestingKit which saw this & saw correct parsing. Also ran against our internal Xcode 10 & Xcode 11 xcresults.