Address inability to export files with bash sensitive characters in filename (#18)
Change Description: These changes make it so that we escape various characters that are sensitive in bash when contained with double-quotes. This should allow exporting files like the debug descriptions without seeing bash bash us about unknown commands.

Test Plan/Testing Performed: Tested with these changes that an XCResult with a text file with a name containing backtick and double quotes could be successfully exported.