Fix the naive approach to consolidating progress bars (#24)
Change Description: These changes make it so that the progress bar consolidation is much more robust than before. Depending on the options you passed, it was possible to see progress bars duplicated. This is now eliminated by not exporting attachments until we've found all of them and determined which URLs they will be exported to. We have a dictionary that uses the URL's path as the key & add attachments that want to be exported to that URL into the corresponding value in the dictionary. We also sort the keys of the dictionary at the end so that the progress bars will be nicely ordered. Display name logic has been simplified to just be based off of the relative name as compared to the base screenshot URL so that new folder additions will automatically get displayName support.

Test Plan/Testing Performed: Tested with these changes that if I provide various options, I only see one progress bar with the same display name. When there is no relevant display name, "Exporting Screenshots" is shown. Ran with xcresults we have locally for Xcode 11 with test plan configs.