Remove interactive mode (#28)
Change Description: We're now at a point where interactive mode is currently more work to keep around than it is likely worth. As part of version 1, we're going to remove the mode now that we allow for multiple options & a highly usable --help command for the various sub-commands and compatible options.

Test Plan/Testing Performed: Tested that with these changes, calling the old interactive mode now leads to the help usage being printed.