Add --test-plan-config flag & deprecate --test-run (#36)
Change Description: The --test-run flag named itself off of the internal implementation of xcresult's ActionTestPlanRunSummary which is used to represent the test plan configuration runs. However, this name is not that transparent to users of xcparse unless theyre familiar with the xcresult internals. Due to this, we should have the flag use the name users who only see the XCResult from within Xcode would expect & there Xcode always refers to it as "Configurations" of the test plan so "--test-plan-config" seems most appropriate to convey that we'll split on each test plan configuration you have in the xcresult.

Test Plan/Testing Performed: Ran the new flag name and was I got the correct output. Ran with the deprecated flag & ensured I got a warning message, but still got exactly what I'd expect.