xcparse should give better error when xcresult path leads to non-xcresult (#47)
Change Description: This fixes #40 by putting an explicit check to see if the xcresult path exists as a directory with an Info.plist before attempting any xcresulttool attempts. If the path does not pass, an error is given to the user highlighting the input path they gave that does not appear to be an xcresult.

Test Plan/Testing Performed: Added unit tests to the commands to confirm the error will get shown.