Arrays of XCResultValueTypes do not parse correctly (#52)
Change Description: This change addresses an issue where arrays (such as [Double] in the performance metrics object) were not properly being decoded. This was due to the need to decode into the intermediate type of XCResultValueType in order to get to the true values underneath.

Test Plan/Testing Performed: Added the xcresult that @OdNairy gave in #51. Added some explicit tests to double-check that the values returned match what we would expect. Also did a quick test if we removed one of the enum cases like "CPU Time" that the metric type enum would properly handle new metric identifier values from new xcresult versions.