Add xcresult 3.26 support (#53)
Change Description: This fixes #50 by adding in the xcresult 3.26 data structures. It also adds back in the userInfo parsing that we removed early on in xcparse creation. This should allow for folks to get info such as the display scale from screenshot attachments. I've made a separate folder for the formatDescriptions so that we can keep track what we've added & make it easier for devs to compare and update when new Xcodes come out.

Test Plan/Testing Performed: Ran the existing unit tests to confirm that parsing is not regressed on them. Need to still add some Xcode 12 xcresults to double-check the new parsing.