[BUG] Xcparse not installing with Xcode 13.0 Beta 4 (#63)
Change Description: These changes fix #62. Swift 5.5 adds async/await & the Swift Package Manager 0.5.0 had an unfortunately named "await" function. The compiler did not take kindly to this & a fix is required in the SPM code as seen done here.

In addition, SPMUtility was transitioned to TSCUtility in a new repo of Swift Tools Core Support in late 2019/early 2020. Information about how to transition could be found in this PR American Express did & this updated blog post by @rderik.

This PR modifies Basic uses to TSCBasic & SPMUtility to TSCUtility from the new repo and pegs us to the current version as of July 2021 that has the commit fixing the await to tsc_await. It should be noted, we still need to go back and clean up all the new warnings as some of our usage from SPM 0.5.0 are deprecated in TSCUtility's latest version, but the functionality appears to work & unit tests pass so that can be a further cleanup task for @rsukumar-cpi

Test Plan/Testing Performed: Ran the build on Xcode 13 beta 4 & confirmed that I could successfully build with the changes. Ran the unit tests & saw success.