Switch over to using an ephemeral URL session for the version check
We're noticing that our CI/CD is getting the error "volume does not support data protection, stripping SQLITE_OPEN_FILEPROTECTION_* flags". Reports from others online point to the URL session attempting to write data that causes this in that particular environment where data protection is unavailable. As we don't need anything to be stored to disk for this, changing to ephemeral will prevent URL session from attempting the write and causing that log warning to be emitted.